pkg/blobserver/s3: fix tests
The storage tets for s3 have been broken since
https://perkeep-review.googlesource.com/c/perkeep/+/15846.

The introduction of a proxycache made a type assertion in the test
panic, and thus the tests would reliably instantly failed if they
weren't skipped.

The simpleist fix for now is to get them back to their previous passing
state by disabling the proxycache.

Change-Id: Ib75b5eabd36decebd8785ba50d4b979a7a0db653